Revert breaking changes of RestClient
